requirements.txt
httpx==0.28.1
anthropic==0.52.0
python-dotenv==1.1.0
pydantic==2.11.3
pykrx==1.0.50
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
apscheduler==3.11.0
aiosqlite==0.21.0
sqlalchemy==2.0.40
greenlet==3.2.3
python-multipart==0.0.20
google-genai==1.14.0
